[refactor] : Performance Schedule 리팩토링 및 수정
