+++ src/main/java/com/example/stagemate/domain/performanceSchedule/PerformanceScheduleType.java
@@ -1,4 +1,4 @@
-package com.example.stagemate.domain.performanceSchedules;
+package com.example.stagemate.domain.performanceSchedule;
 
 public enum PerformanceScheduleType {
     START,
