+++ src/main/java/com/example/stagemate/repository/PerformanceScheduleRepository.java
@@ -1,7 +1,7 @@
 package com.example.stagemate.repository;
 
-import com.example.stagemate.domain.performanceSchedules.PerformanceSchedule;
-import com.example.stagemate.domain.performanceSchedules.PerformanceScheduleType;
+import com.example.stagemate.domain.performanceSchedule.PerformanceSchedule;
+import com.example.stagemate.domain.performanceSchedule.PerformanceScheduleType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
@@ -9,8 +9,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 public interface PerformanceScheduleRepository extends JpaRepository<PerformanceSchedule, Long> {
-    List<PerformanceSchedule> findByPerformance_PerformanceIdIn(List<Long> performanceIds);
-    PerformanceSchedule findByPerformance_PerformanceIdAndPerformanceScheduleType(Long performanceId, PerformanceScheduleType type);
+    PerformanceSchedule findByPerformance_IdAndPerformanceScheduleType(Long performanceId, PerformanceScheduleType type);
     List<PerformanceSchedule> findByScheduleDate(LocalDate scheduleDate);
 
     @Query("SELECT ps FROM PerformanceSchedule ps JOIN FETCH ps.performance WHERE ps.scheduleDate BETWEEN :startDate AND :endDate")
